pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
flask = "^3.0.0"
sqlalchemy = "2.0.27"
psycopg2-binary = "2.9.9"
requests = "2.31.0"
pandas = "2.2.0"
extraction = "^0.3"
loading = "^0.0.1"
numpy = "1.26.4"
etl = "^0.0.1"
pyarrow = "15.0.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"